% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p_value.R
\name{calculate_p_value}
\alias{calculate_p_value}
\title{Calculate P-value for Contingency Table}
\usage{
calculate_p_value(df, test)
}
\arguments{
\item{df}{A contingency table in wide format.}

\item{test}{The statistical test to use ("chi2" or "fisher").}
}
\value{
A dataframe with the p-value added as a column.
}
\description{
Calculates the p-value using the selected statistical test (\code{chi2} or \code{fisher}).
}
\keyword{internal}
