% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentages.R
\name{calculate_percentages}
\alias{calculate_percentages}
\title{Calculate Percentages}
\usage{
calculate_percentages(df, out)
}
\arguments{
\item{df}{A dataframe with counts.}

\item{out}{Output type: "obf" for obfuscation, "pct" for standard percentage.}
}
\value{
A dataframe with percentages.
}
\description{
Calculates percentages with obfuscation for small values if \code{out} is set to "obf".
}
\keyword{internal}
