% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency_tab.R
\name{contingency_tab}
\alias{contingency_tab}
\title{contingency_tab}
\usage{
contingency_tab(
  df,
  x,
  y,
  n.col = TRUE,
  na.use = "as.note",
  test = "chi2",
  out = "obf",
  inline.title = FALSE,
  label.title = TRUE,
  lang = Sys.getlocale("LC_CTYPE")
)
}
\arguments{
\item{df}{A dataframe containing categorical data.}

\item{x}{The row variable.}

\item{y}{The column variable.}

\item{n.col}{Display count per column.}

\item{na.use}{Mode of displaying missing values:
\itemize{
\item "as.cat": As a regular category.
\item "as.note": As a note readable by stylize.
\item "no": Do not display missing values.
}}

\item{test}{Statistical test used:
\itemize{
\item "chi2": Chi-square test
\item "fisher": Fisher's exact test
\item "none": No test
}}

\item{out}{Output type:
\itemize{
\item "obf": Percentage with obfuscation for small values
\item "pct": Standard percentage
\item "n": Exact number
}}

\item{inline.title}{Add the row variable title inline in the table}

\item{label.title}{Replace the row variable title with its label}

\item{lang}{Table language ("fr" or "en")}
}
\value{
A proportion dataframe
}
\description{
Generates a simple contingency table.
}
\examples{
warpbreaks |>
contingency_tab(wool, tension)
}
