% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contingency_table.R
\name{create_contingency_table}
\alias{create_contingency_table}
\title{Create Contingency Table}
\usage{
create_contingency_table(df, x, y)
}
\arguments{
\item{df}{A dataframe with categorical data.}

\item{x}{The row variable.}

\item{y}{The column variable.}
}
\value{
A contingency table in wide format.
}
\description{
Generates a contingency table with counts for each combination of \code{x} and \code{y}.
}
\keyword{internal}
