% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_counts_per_column.R
\name{display_counts_per_column}
\alias{display_counts_per_column}
\title{Display Counts per Column}
\usage{
display_counts_per_column(df, y, n.col)
}
\arguments{
\item{df}{A dataframe with categorical data.}

\item{y}{The column variable.}

\item{n.col}{Logical, if \code{TRUE}, display counts per column.}
}
\value{
A dataframe with counts per column level in \code{y}.
}
\description{
Adds count information for each level of \code{y} if \code{n.col} is \code{TRUE}.
}
\keyword{internal}
