% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p_value.R
\name{format_p_value}
\alias{format_p_value}
\title{Format P-value}
\usage{
format_p_value(p)
}
\arguments{
\item{p}{The p-value to format.}
}
\value{
A formatted p-value.
}
\description{
Helper function to format p-value.
}
\keyword{internal}
