% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missing_values.R
\name{handle_missing_values}
\alias{handle_missing_values}
\title{Handle Missing Values in Contingency Table}
\usage{
handle_missing_values(df, x, na.use, lang)
}
\arguments{
\item{df}{A dataframe with categorical data.}

\item{x}{The row variable.}

\item{na.use}{Mode of displaying missing values:
\itemize{
\item \code{"as.cat"}: As a regular category.
\item \code{"as.note"}: Count missing values as a note.
}}

\item{lang}{Table language (\code{"fr"} or \code{"en"})}
}
\value{
A list with a dataframe (possibly with a missing category) and a missing count.
}
\description{
Adjusts for missing values in \code{x} based on the \code{na.use} parameter.
}
\keyword{internal}
