% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_stylize.R
\name{st_stylize}
\alias{st_stylize}
\title{Stylize}
\usage{
st_stylize(data, theme = "classy")
}
\arguments{
\item{data}{A table or dataframe to be formatted.}

\item{theme}{Flextable theme you want to use}
}
\value{
A Flextable object ready to be printed in your Word document.
}
\description{
Handles the table layout by organizing its structure, formatting cells, and aligning content for optimal readability.
}
\examples{
# Example of using the st_stylize function

# Create a small dataset
example_data <- data.frame(
  Name = c("Alice", "Bob", "Charlie"),
  Age = c(25, 30, 35),
  Profession = c("Engineer", "Doctor", "Teacher")
)

# Apply the st_stylize function to the dataset
st_stylized_table <- st_stylize(example_data)

# Display the result (here, st_stylized_table is the formatted Flextable object)
print(st_stylized_table)
}
