% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_utils_add_custom_note.R
\name{st_utils_add_custom_note}
\alias{st_utils_add_custom_note}
\title{Add Custom Notes to Flextable}
\usage{
st_utils_add_custom_note(ft, data, n)
}
\arguments{
\item{ft}{A flextable object to which the note will be added.}

\item{data}{A dataframe containing \verb{<note:...>} tags in the \code{Variable} column.}

\item{n}{Index number for the footnote.}
}
\value{
A flextable object with the custom notes added.
}
\description{
Detects \verb{<note:...>} tags in the \code{Variable} column of \code{data} and adds a custom footnote to the flextable.
}
\keyword{internal}
