% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_utils_add_missing.R
\name{st_utils_add_missing}
\alias{st_utils_add_missing}
\title{Add Missing Data Information to Flextable with Language Support}
\usage{
st_utils_add_missing(ft, data, lang = Sys.getlocale("LC_CTYPE"))
}
\arguments{
\item{ft}{A flextable object to which missing data information will be added.}

\item{data}{A dataframe containing a \code{Variable} column with \verb{<missing:X>} tags for missing data counts.}

\item{lang}{Language for the footer text. Accepts \code{"en"} for English or \code{"fr"} for French.}
}
\value{
A flextable object with a summary note added in the footer.
}
\description{
This internal utility function detects and processes \verb{<missing:X>} tags in the \code{Variable} column of \code{data},
removing the tags from the main content and appending a summary note to the flextable footer, in the specified language.
}
\details{
\itemize{
\item The function detects \verb{<missing:X>} tags in the \code{Variable} column of \code{data}, where \code{X} is the count of missing data.
\item These tags are removed from the main content, and a summary of missing data is generated.
\item The summary note, detailing missing counts for each variable or for multiple groups, is appended to the footer of \code{ft} in the specified language.
}
}
\examples{
# (For internal use only; no example provided.)
}
\keyword{internal}
