% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_utils_add_nonadd_note.R
\name{st_utils_add_nonadd_note}
\alias{st_utils_add_nonadd_note}
\title{Add Non-Additivity Note to Flextable}
\usage{
st_utils_add_nonadd_note(ft, data, lang, n)
}
\arguments{
\item{ft}{A flextable object to which the note will be added.}

\item{data}{A dataframe containing \verb{<nonadd>} tags in the \code{Variable} column.}

\item{lang}{Language for the note text. Accepts \code{"en"} for English or \code{"fr"} for French.}

\item{n}{Index number for the footnote.}
}
\value{
A flextable object with the non-additivity note added.
}
\description{
Adds a footnote for non-additive categories in the \code{Variable} column.
}
\keyword{internal}
