% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_contingency.R
\name{tab_contingency}
\alias{tab_contingency}
\title{tab_contingency}
\usage{
tab_contingency(
  df,
  x,
  y,
  n.col = TRUE,
  na.use = "as.note",
  test = "chi2",
  out = "obf",
  inline.title = FALSE,
  label.title = TRUE,
  lang = Sys.getlocale("LC_CTYPE")
)
}
\arguments{
\item{df}{A dataframe containing categorical data.}

\item{x}{The row variable.}

\item{y}{The column variable.}

\item{n.col}{Display count per column.}

\item{na.use}{Mode of displaying missing values:
\itemize{
\item "as.cat": As a regular category.
\item "as.note": As a note readable by st_stylize.
\item "no": Do not display missing values.
}}

\item{test}{Statistical test used:
\itemize{
\item "chi2": Chi-square test
\item "fisher": Fisher's exact test
\item "none": No test
}}

\item{out}{Output type:
\itemize{
\item "obf": Percentage with obfuscation for small values
\item "pct": Standard percentage
\item "n": Exact number
}}

\item{inline.title}{Add the row variable title inline in the table}

\item{label.title}{Replace the row variable title with its label}

\item{lang}{Table language ("fr" or "en")}
}
\value{
A proportion dataframe
}
\description{
Generates a simple contingency table.
}
\examples{
# Example 1: Simple contingency table without advanced options
tab_contingency(warpbreaks, wool, tension)

# Example 2: Treat missing values in `wool` as a separate category
warpbreaks_with_na <- warpbreaks
warpbreaks_with_na$wool[c(5, 15, 25)] <- NA  # Add some NA values for demonstration
tab_contingency(warpbreaks_with_na, wool, tension, na.use = "as.cat")

# Example 3: Display missing values as a note
tab_contingency(warpbreaks_with_na, wool, tension, na.use = "as.note")

# Example 4: Contingency table with chi-square test
tab_contingency(warpbreaks, wool, tension, test = "chi2")

# Example 5: Contingency table with Fisher's exact test for small sample sizes
tab_contingency(warpbreaks, wool, tension, test = "fisher")

# Example 6: Display percentages without obfuscation
tab_contingency(warpbreaks, wool, tension, out = "pct")

# Example 7: Display percentages with obfuscation for small values
tab_contingency(warpbreaks, wool, tension, out = "obf")

# Example 8: Add inline title for the `wool` variable
tab_contingency(warpbreaks, wool, tension, inline.title = TRUE)

# Example 9: Use a variable label for `wool`
labelled::var_label(warpbreaks$wool) <- "Type de laine"
tab_contingency(warpbreaks, wool, tension, label.title = TRUE)

# Example 10: Combine multiple options
tab_contingency(warpbreaks_with_na, wool, tension,
                n.col = TRUE, na.use = "as.cat", test = "chi2", out = "pct",
                inline.title = TRUE, label.title = TRUE)
}
