% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_contingency_binded.R
\name{tab_contingency_binded}
\alias{tab_contingency_binded}
\title{tab_contingency_binded}
\usage{
tab_contingency_binded(
  df,
  ...,
  y,
  p_group = NULL,
  n.col = TRUE,
  na.use = "as.note",
  test = "chi2",
  out = "obf",
  label.title = TRUE,
  lang = Sys.getlocale("LC_CTYPE")
)
}
\arguments{
\item{df}{A dataframe containing categorical data.}

\item{...}{One or more row variable.}

\item{y}{The column variable.}

\item{p_group}{An integer vector representing the width of each column group.}

\item{n.col}{Display count per column.}

\item{na.use}{Mode of displaying missing values:
\itemize{
\item "as.cat": As a regular category.
\item "as.note": As a note readable by st_stylize.
\item "no": Do not display missing values.
}}

\item{test}{Statistical test used:
\itemize{
\item "chi2": Chi-square test
\item "fisher": Fisher's exact test
\item "none": No test
}}

\item{out}{Output type:
\itemize{
\item "obf": Percentage with obfuscation for small values
\item "pct": Standard percentage
\item "n": Exact number
}}

\item{label.title}{Replace the row variable title with its label}

\item{lang}{Table language ("fr" or "en")}
}
\value{
A proportion dataframe
}
\description{
Bind multiple contingency tables with or without grouped columns
}
\examples{
warpbreaks$breaks <- warpbreaks$breaks > 20
# Example 1: Bind contingency tables without column grouping
tab_contingency_binded(warpbreaks, wool, tension, y = breaks, test = "chi2")

# Example 2: Bind contingency tables with grouped columns (grouping sizes 2 and 1)
tab_contingency_binded(warpbreaks, wool, breaks, y = tension, p_group = c(2, 1), na.use = "as.note")

# Note: For additional parameter information, see ?tab_contingency or ?tab_contingency_grouped.
}
