% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_contingency_grouped.R
\name{tab_contingency_grouped}
\alias{tab_contingency_grouped}
\title{tab_contingency_grouped}
\usage{
tab_contingency_grouped(
  df,
  x,
  y,
  p_group,
  n.col = TRUE,
  na.use = "as.note",
  test = "chi2",
  out = "obf",
  inline.title = FALSE,
  label.title = TRUE,
  lang = Sys.getlocale("LC_CTYPE")
)
}
\arguments{
\item{df}{A dataframe containing categorical data.}

\item{x}{The row variable.}

\item{y}{The column variable.}

\item{p_group}{An integer vector representing the width of each column group.}

\item{n.col}{Display count per column.}

\item{na.use}{Mode of displaying missing values:
\itemize{
\item "as.cat": As a regular category.
\item "as.note": As a note readable by st_stylize.
\item "no": Do not display missing values.
}}

\item{test}{Statistical test used:
\itemize{
\item "chi2": Chi-square test
\item "fisher": Fisher's exact test
\item "none": No test
}}

\item{out}{Output type:
\itemize{
\item "obf": Percentage with obfuscation for small values
\item "pct": Standard percentage
\item "n": Exact number
}}

\item{inline.title}{Add the row variable title inline in the table}

\item{label.title}{Replace the row variable title with its label}

\item{lang}{Table language ("fr" or "en")}
}
\value{
A proportion dataframe
}
\description{
Generates a contingency table with grouped columns
}
\examples{
# Basic usage: grouped contingency table with specified column groups
# For more details on additional parameters, see ?tab_contingency

# Example 1: Group columns into two groups with sizes 2 and 1
tab_contingency_grouped(warpbreaks, wool, tension, c(2, 1))

# Example 2: Group columns with different group sizes, specifying missing
# values handling and a statistical test
warpbreaks_with_na <- warpbreaks
warpbreaks_with_na$wool[c(5, 15, 25)] <- NA
tab_contingency_grouped(warpbreaks_with_na, wool, tension,
                        c(1, 2), na.use = "as.note", test = "chi2")

# Example 3: Group columns with percentage output and labels inline
tab_contingency_grouped(warpbreaks, wool, tension,
                        c(2, 1), out = "pct", inline.title = TRUE)

# Example 4: Grouped table with column label replacement for `wool`
labelled::var_label(warpbreaks$wool) <- "Type de laine"
tab_contingency_grouped(warpbreaks, wool, tension,
                        c(1, 2), label.title = TRUE)

# Note: Additional customization options are available.
# Refer to ?tab_contingency for more details on parameters such as `n.col`,
# `na.use`, `test`, and `out`.
}
