% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_contingency_yn.R
\name{tab_contingency_yn}
\alias{tab_contingency_yn}
\title{tab_contingency_yn}
\usage{
tab_contingency_yn(
  df,
  x,
  y,
  p_group = NULL,
  pos.value = TRUE,
  n.col = TRUE,
  na.use = "as.note",
  test = "chi2",
  out = "obf",
  inline.title = FALSE,
  label.title = TRUE,
  lang = Sys.getlocale("LC_CTYPE")
)
}
\arguments{
\item{df}{A dataframe containing categorical data.}

\item{x}{The binary row variable to focus on (e.g., yes/no, true/false).}

\item{y}{The column variable.}

\item{p_group}{An integer vector representing the width of each column group. If \code{NULL}, columns are not grouped.}

\item{pos.value}{The value of \code{x} to retain in the final table (e.g., \code{TRUE} for focusing on positive cases).}

\item{n.col}{Display count per column.}

\item{na.use}{Mode of displaying missing values:
\itemize{
\item "as.cat": As a regular category.
\item "as.note": As a note readable by stylization functions.
\item "no": Do not display missing values.
}}

\item{test}{Statistical test used:
\itemize{
\item "chi2": Chi-square test.
\item "fisher": Fisher's exact test.
\item "none": No test.
}}

\item{out}{Output type:
\itemize{
\item "obf": Percentage with obfuscation for small values.
\item "pct": Standard percentage.
\item "n": Exact number.
}}

\item{inline.title}{Add the row variable title inline in the table.}

\item{label.title}{Replace the row variable title with its label if a label exists.}

\item{lang}{Table language ("fr" or "en").}
}
\value{
A dataframe containing the filtered contingency table for the specified "positive" value of \code{x}.
}
\description{
Generates a contingency table for a binary variable, filtering to focus on a specified "positive" value, with optional column grouping.
}
\examples{
# Example 1: Basic usage without column grouping, focusing on TRUE values
tab_contingency_yn(warpbreaks, wool, tension, pos.value = "A")

# Example 2: With grouped columns, focusing on a positive value with Chi-square test
tab_contingency_yn(warpbreaks, wool, tension, p_group = c(2, 1), pos.value = "A", test = "chi2")

# Note: Refer to ?tab_contingency and ?tab_contingency_grouped for further parameter details.
}
