% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_contingency_yn_binded.R
\name{tab_contingency_yn_binded}
\alias{tab_contingency_yn_binded}
\title{tab_contingency_yn_binded}
\usage{
tab_contingency_yn_binded(
  df,
  ...,
  y,
  p_group = NULL,
  title,
  pos.value = TRUE,
  n.col = TRUE,
  na.use = "as.note",
  test = "chi2",
  out = "obf",
  label.title = TRUE,
  lang = Sys.getlocale("LC_CTYPE")
)
}
\arguments{
\item{df}{A dataframe containing categorical data.}

\item{...}{One or more binary row variables to include in the combined table.}

\item{y}{The column variable.}

\item{p_group}{An integer vector representing the width of each column group. If \code{NULL}, columns are not grouped.}

\item{title}{A title row added at the top of the final combined table.}

\item{pos.value}{Value of ... to consider as the positive value.}

\item{n.col}{Display count per column.}

\item{na.use}{Mode of displaying missing values:
\itemize{
\item "as.cat": As a regular category.
\item "as.note": As a note readable by stylization functions.
\item "no": Do not display missing values.
}}

\item{test}{Statistical test used:
\itemize{
\item "chi2": Chi-square test.
\item "fisher": Fisher's exact test.
\item "none": No test.
}}

\item{out}{Output type:
\itemize{
\item "obf": Percentage with obfuscation for small values.
\item "pct": Standard percentage.
\item "n": Exact number.
}}

\item{label.title}{Replace the row variable title with its label if a label exists.}

\item{lang}{Table language ("fr" or "en").}
}
\value{
A dataframe containing the combined contingency tables for the specified "positive" values of each binary variable in \code{...}.
}
\description{
Combines multiple contingency tables focused on a specific "positive" value, with optional column grouping, into a single table.
}
\examples{
warpbreaks$breaks <- ifelse(warpbreaks$breaks > 20, "A", "B")
# Example 1: Combine contingency tables without column grouping
tab_contingency_yn_binded(warpbreaks, breaks, wool, y = tension,
                          title = "Combined Wool and Tension Table", pos.value = "A")

# Example 2: Combine contingency tables with grouped columns
tab_contingency_yn_binded(warpbreaks, breaks, wool, y = tension, p_group = c(2,1),
                          title = "Combined Wool and Tension Table", pos.value = "A")

# Note: Refer to ?tab_prop_yn for additional parameter details.
}
