% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_utils_add_inline_title.R
\name{tab_utils_add_inline_title}
\alias{tab_utils_add_inline_title}
\title{Add Inline Title to Contingency Table}
\usage{
tab_utils_add_inline_title(df, x, df_init, inline.title, label.title, na_count)
}
\arguments{
\item{df}{A contingency table.}

\item{x}{The row variable.}

\item{df_init}{The initial dataframe to retrieve variable labels.}

\item{inline.title}{Logical, if \code{TRUE}, add title inline.}

\item{label.title}{Logical, if \code{TRUE}, replace row variable title with its label.}

\item{na_count}{Count of missing values if \code{na.use} is "as.note".}
}
\value{
A dataframe with the inline title if specified.
}
\description{
Adds the row variable title inline in the table if \code{inline.title} is \code{TRUE}.
}
\keyword{internal}
