% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_utils_handle_na_group.R
\name{tab_utils_handle_na_group}
\alias{tab_utils_handle_na_group}
\title{Handle Missing Values for Grouped Table}
\usage{
tab_utils_handle_na_group(df_list, na.use, inline.title)
}
\arguments{
\item{df_list}{A list of dataframes.}

\item{na.use}{Mode for displaying missing values.}

\item{inline.title}{Add the row variable title inline.}
}
\value{
A list of dataframes with adjusted missing values display.
}
\description{
Adjusts columns and row names in the dataframe list to handle missing values.
}
\keyword{internal}
