% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_utils_percent.R
\name{tab_utils_percent}
\alias{tab_utils_percent}
\title{Calculate Percentages}
\usage{
tab_utils_percent(df, out)
}
\arguments{
\item{df}{A dataframe with counts.}

\item{out}{Output type: "obf" for obfuscation, "pct" for standard percentage.}
}
\value{
A dataframe with percentages.
}
\description{
Calculates percentages with obfuscation for small values if \code{out} is set to "obf".
}
\keyword{internal}
