% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_utils_pval.R
\name{tab_utils_pval}
\alias{tab_utils_pval}
\title{Calculate P-value for Contingency Table}
\usage{
tab_utils_pval(df, test)
}
\arguments{
\item{df}{A contingency table in wide format.}

\item{test}{The statistical test to use ("chi2" or "fisher").}
}
\value{
A dataframe with the p-value added as a column.
}
\description{
Calculates the p-value using the selected statistical test (\code{chi2} or \code{fisher}).
}
\keyword{internal}
