% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_utils_split_col.R
\name{tab_utils_split_col}
\alias{tab_utils_split_col}
\title{Split Columns into Groups}
\usage{
tab_utils_split_col(cols, p_group)
}
\arguments{
\item{cols}{A vector of column names to be split.}

\item{p_group}{An integer vector specifying the size of each group.}
}
\value{
A list of column name groups.
}
\description{
This function divides a vector of columns into specified group sizes.
}
