% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_utils_yn_title.R
\name{tab_utils_yn_title}
\alias{tab_utils_yn_title}
\title{tab_utils_yn_title}
\usage{
tab_utils_yn_title(
  df,
  df_init,
  x,
  label.title = TRUE,
  na_count = NULL,
  inline.title = FALSE
)
}
\arguments{
\item{df}{The dataframe containing the contingency table.}

\item{df_init}{The initial dataframe.}

\item{x}{The variable for which to generate a title.}

\item{label.title}{Logical; if \code{TRUE}, replace the row variable title with its label if available.}

\item{na_count}{A string indicating the missing value count to append to the title.}

\item{inline.title}{Logical; if \code{TRUE}, places the title inline within the table instead of as the column name.}
}
\value{
The dataframe with the title appropriately set as the column name or inline.
}
\description{
Generates and applies the title for a contingency table, using the label of the variable if it exists.
If there are missing values and \code{na.use} is set to "as.note", it appends a missing count note.
It also handles inline title placement or column name replacement.
}
\keyword{internal}
