% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_classy.R
\name{theme_classy}
\alias{theme_classy}
\title{Classy Theme}
\usage{
theme_classy(ft)
}
\arguments{
\item{ft}{A \code{flextable} object that you want to style.}
}
\value{
A \code{flextable} object with the applied classy theme formatting.
}
\description{
Applies a refined and professional style to a \code{flextable} object, formatting the table headers, body, and footer for improved readability and visual appeal.
}
\examples{
# Example usage of classy_theme*
library(flextable)
library(officer)
ft <- flextable(head(iris))  # Create a simple flextable
styled_ft <- theme_classy(ft)  # Apply the classy theme to the flextable
styled_ft
}
